% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_functions.R
\name{cdAFLAPI.update}
\alias{cdAFLAPI.update}
\title{Update Package Version}
\usage{
cdAFLAPI.update(...)
}
\description{
Function to install the most recent version of cdAFLAPI. Will prompt for a session restart if using R Studio (recommended).
}
\examples{
cdAFLAPI.update()
}
