% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_functions.R
\name{cdAFLAPI.version}
\alias{cdAFLAPI.version}
\title{Current Version of cdAFLAPI}
\usage{
cdAFLAPI.version(info = "Version")
}
\arguments{
\item{info}{A string for which information to return.
\code{info='Version'} will return a string of the current version number. (Default)
\code{info='Details'} will return a list containing package metadata of your current version.}
}
\description{
Return version details of the current version of the cdAFLAPI your system is running.
}
\examples{
cdAFLAPI.version()
}
