% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cdAPI}
\alias{cdAPI}
\title{Champion Data AFL API}
\usage{
cdAPI(
  endpoint,
  envir = c("", "sandbox", "dev"),
  base_url = "https://api.afl.championdata.io",
  version = "v1",
  df = TRUE
)
}
\arguments{
\item{endpoint}{The full path of an endpoint including path and query parameters, excluding the base_url and version components. eg: "Leagues" rather than "https://api.afl.championdata.io/v1/Leagues"}

\item{envir}{A short code representing the environment to hit. Empty string = Production, 'sandbox' = Sandbox environment, 'dev' = Development environment (Internal Champion Data use only).}

\item{base_url}{The base URL for the API. Defaults to the AFL API.}

\item{version}{The version number of the API. Defaults to 1.}

\item{df}{Logical. Convert to data.frame? Note: Will only work for endpoints with a flat JSON response.}
}
\value{
A numerical value for league ID.
}
\description{
Hit the Champion Data AFL API with a given endpoint.
}
\examples{
cdAPI('matches/113770401')
}
