% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cdAPIPOSTresponse}
\alias{cdAPIPOSTresponse}
\title{Champion Data AFL API Response (POST Endpoint)}
\usage{
cdAPIPOSTresponse(
  matchId,
  endpoint,
  payload,
  baseUrl = "https://api.afl.championdata.io",
  version = "v1",
  envir = ""
)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{endpoint}{A (case-insensitive) string detailing which endpoint to hit (either 'player' or 'squad').}

\item{payload}{A nested list containing the parcels of data to be returned by the API. See: \code{createPayload()} as a method of simplifying this step.}

\item{version}{The version number of the API. Defaults to 1.}

\item{envir}{A short code representing the environment to hit. Empty string = Production, 'sandbox' = Sandbox environment, 'dev' = Development environment (Internal Champion Data use only).}

\item{base_url}{The base URL for the API. Defaults to the AFL API.}
}
\value{
A response object from the API call.
}
\description{
Hit either the Players or Squads Champion Data AFL API via the POST Endpoint.
}
\examples{
cdAPIPOSTresponse(matchId = 120390401, endpoint = "player", payload = payload)
}
