% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cdAPIresponse}
\alias{cdAPIresponse}
\title{Champion Data AFL API Response}
\usage{
cdAPIresponse(
  matchId,
  endpoint,
  envir = c("", "sandbox", "dev"),
  base_url = "https://api.afl.championdata.io",
  version = "v1"
)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match (only required for match-level endpoints.)}

\item{endpoint}{The full path of an endpoint including path and query parameters, excluding the base_url and version components. eg: "Leagues" rather than "https://api.afl.championdata.io/v1/Leagues"}

\item{envir}{A short code representing the environment to hit. Empty string = Production, 'sandbox' = Sandbox environment, 'dev' = Development environment (Internal Champion Data use only).}

\item{base_url}{The base URL for the API. Defaults to the AFL API.}

\item{version}{The version number of the API. Defaults to 1.}

\item{verbose}{Logical, default is \code{FALSE}. If \code{TRUE}, prints the full URL string passed to the API.}
}
\value{
A response object.
}
\description{
Hit the Champion Data AFL API with a given endpoint to retrieve a response object.
}
\examples{
cdAPIresponse(matchId,paste('matches',matchId,'entries',sep='/')
}
