% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{createPayLoad}
\alias{createPayLoad}
\title{createPayload}
\usage{
createPayLoad(endpoint, ...)
}
\arguments{
\item{endpoint}{A string of either "player" or "squad" dictating the POST endpoint the payload is being built for.}

\item{list(s)}{Pass in any number of single lists to be combined into one payload.}
}
\value{
A nested list in the correct structure to pass into payload arguments of \code{getPlayerStatsPOST()} and \code{getSquadStatsPOST()}
}
\description{
Create a payload for \code{getPlayerStatsPOST()} and \code{getSquadStatsPOST()} functions.
}
\examples{
D50_Tackles_Q1 = list(metricCodes = list("TACKLE"), zones = list('D50'), periods = list(1), id = "D50_Tackles_Q1")
createPayLoad(endpoint = "squad", D50_Tackles_Q1)
}
