% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{extractError}
\alias{extractError}
\title{extractError}
\usage{
extractError(response)
}
\value{
An error message (if any) provided from the API.
}
\description{
This function is used inside of the new httr2 version of cdAPIResponse() to extract the error message from the response object if the response is not successful.
}
\keyword{internal}
