% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fetchAllLeagues}
\alias{fetchAllLeagues}
\title{fetchAllLeagues}
\usage{
fetchAllLeagues(...)
}
\value{
A DF with all info from /leagues endpoint
}
\description{
Helper function for inside getLeagueLevelsInfo(): used to hit /leagues endpoint for all leagues
}
\keyword{internal}
