% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fetchLeagueLevelCombos}
\alias{fetchLeagueLevelCombos}
\title{fetchLeagueLevelCombos}
\usage{
fetchLeagueLevelCombos(leagueId, levelId, ...)
}
\value{
A DF with info for a given leagueId & levelId
}
\description{
Helper function for inside getLeagueLevelsInfo(): used to get the league-level info for a given league-level combo
}
\keyword{internal}
