% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fetchLevels}
\alias{fetchLevels}
\title{fetchLevels}
\usage{
fetchLevels(leagueId, ...)
}
\value{
A DF with level info for a given leagueId
}
\description{
Helper function for inside getLeagueLevelsInfo(): used to get all the levels for a given leagueId
}
\keyword{internal}
