% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{flattenPaginatedTrx}
\alias{flattenPaginatedTrx}
\title{flattenPaginatedTrx}
\usage{
flattenPaginatedTrx(resp)
}
\value{
A flattened DF from a page of API response.
}
\description{
Helper function used in getStatisticFlow() to take a page of API response and flatten it into a single DF
}
\keyword{internal}
