% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getAFLClubTrxFeed}
\alias{getAFLClubTrxFeed}
\title{AFL Club Feed Transaction File}
\usage{
getAFLClubTrxFeed(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame in the same format and structure as the AFL Club Transaction Feed file formerly available via the support site.
\itemize{
   \item \code{MATCH_ID} A unique numerical identifier of a match.
   \item \code{MATCH_DATE}The calendar date of a match in local time.
   \item \code{MATCH_TIME} The local start time of a match.
   \item \code{SEASON_ID} A numerical identifier of a season.
   \item \code{GROUP_ROUND_NO} The round number of the match. Continues to count up during finals.
   \item \code{VENUE_NAME} The name of the match venue.
   \item \code{HOME_SQUAD} The name of the home squad.
   \item \code{HOME_SCORE} The score of the home squad.
   \item \code{AWAY_SQUAD} The name of the away squad.
   \item \code{AWAY_SCORE} The score of the away squad.
   \item \code{MATCH_TRX_ID} A unique numerical identifier for a given transaction, used for ordering chronologically.
   \item \code{SEQUENCE} Unique identifier of events within a single transaction. Used for multi-player transactions.
   \item \code{PERIOD} The period (quarter) of the match.
   \item \code{PERIOD_SECS} The elapsed time in seconds for live periods, or the period length for completed periods.
   \item \code{STATISTIC_CODE} The statistic code for a given transaction.
   \item \code{PERSON_ID} A unique numerical identifier for the player assigned to the transaction (if applicable).
   \item \code{FULLNAME} The fullname of the player assigned to the transaction (if applicable).
   \item \code{SQUAD_NAME} The name of the squad assigned to the transaction.
   \item \code{OPP_SQUAD} The opposing squad to the one assigned to the transaction.
   \item \code{AR_ID} Away Ruck PERSON_ID (if applicable).
   \item \code{AR} Away Ruck FULLNAME (if applicable).
   \item \code{H1_ID} Home Midfielder 1 PERSON_ID (if applicable).
   \item \code{H1} Home Midfielder 1 FULLNAME (if applicable).
   \item \code{H2_ID} Home Midfielder 2 PERSON_ID (if applicable).
   \item \code{H2} Home Midfielder 2 FULLNAME (if applicable).
   \item \code{H3_ID} Home Midfielder 3 PERSON_ID (if applicable).
   \item \code{H3} Home Midfielder 3 FULLNAME (if applicable).
   \item \code{A1_ID} Away Midfielder 1 PERSON_ID (if applicable).
   \item \code{A1} Away Midfielder 1 FULLNAME (if applicable).
   \item \code{A2_ID} Away Midfielder 2 PERSON_ID (if applicable).
   \item \code{A2} Away Midfielder 2 FULLNAME (if applicable).
   \item \code{A3_ID} Away Midfielder 3 FULLNAME (if applicable).
   \item \code{A3} Away Midfielder 3 FULLNAME (if applicable).
   \item \code{ZONE_LOGICAL_AFL} The zone in which the transaction takes place, relative to the squad in possession.
   \item \code{ZONE_PHYSICAL_AFL} The zone in which the transaction takes place, relative to the physical ground.
   \item \code{TRUEX} The x coordinate of the transaction.
   \item \code{TRUEY} The y coordinate of the transaction.
   \item \code{VENUE_LENGTH} The length of the venue in metres.
   \item \code{VENUE_WIDTH} The width of the venue in metres.
   \item \code{STDX} The x coordinate of the transaction transformed to a standard ground dimension (160x141).
   \item \code{STDY} The y coordinate of the transaction transformed to a standard ground dimension (160x141).
   \item \code{XY_FLIP}
   \item \code{INITIAL_TRX_ID} The transaction ID of the start of the chain.
   \item \code{FINAL_TRX_ID} The transaction ID of the end of the chain.
   \item \code{CHAIN_SQUAD} The name of the squad in possession during the chain.
   \item \code{INITIAL_STATE} The starting state of the chain.
   \item \code{FINAL_STATE} The final state of the chain.
   \item \code{ZONE_LOGICAL_INITIAL} The zone in which the first transaction of the chain takes place, relative to the squad in possession.
   \item \code{FINAL_ZONE_LOGICAL} The zone in which the final transaction of the chain takes place, relative to the squad in possession.
   \item \code{LAUNCH_PERSON_ID} A unique numerical identifier of the player who launched the chain.
   \item \code{LAUNCH_PLAYER} The full name of the player who launched the chain.
   \item \code{GUILTY_PERSON_ID} Turnover player PERSON_ID (if applicable).
   \item \code{GUILTY_PLAYER} Turnover player FULLNAME (if applicable).
   \item \code{PARAM1} Additional captured parameters on transactions, showing the Shot Source, Free Kick Reason, Inside 50 Side or Kick In Direction
   \item \code{PARAM2} Additional captured parameters on transactions, showing the Shot Type, Free Kick Context or Inside 50 Type
   \item \code{PARAM3} Additional captured parameters on transactions, showing the Shot Angle
   \item \code{PARAM4} Additional captured parameters on transactions, showing the Shot Distance
   \item \code{KICK_FOOT} On kicking transactions, the foot of the kicking player.
   \item \code{KICK_INTENT} On kicking transactions, the intent of the kicking player.
   \item \code{KICK_DISTANCE} On kicking transactions, a description of the kicks distance.
   \item \code{KICK_DIRECTION} On kicking transactions, a description of the kicks direction.
   \item \code{PRESSURE_LEVEL} The name of the type of pressure applied (ie. Set Position).
   \item \code{PRESSURE_PLAYER_ID} A unique numerical identifier for the first pressure player.
   \item \code{PRESSURE_PLAYER} The fullname of the first pressure player.
   \item \code{PRESSURE_PLAYER2_ID} A unique numerical identifier for the second pressure player.
   \item \code{PRESSURE_PLAYER2} The fullname of the second pressure player.
   \item \code{PRESSURE_POINTS} The points value of the type of pressure applied.
}
}
\description{
Get the AFL Club Trx Feed file formerly available via the support site.
}
\examples{
getAFLClubTrxFeed(matchId = 120390401)
}
