% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getBench}
\alias{getBench}
\title{Match Bench}
\usage{
getBench(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of all players for the match and their current bench status.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{id} A unique numerical identifier of the player.
   \item \code{name} The full name of the player.
   \item \code{display} The display name of the player, represented as first initial and surname.
   \item \code{status} The player's current status.
   \item \code{elapsed} Seconds elapsed within the current period since the player's last interchange move.
}
}
\description{
Get the current bench status of all players in a match.
}
\examples{
getBench(216085122)
}
