% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getChains}
\alias{getChains}
\title{Match Chains}
\usage{
getChains(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of possession chains for a match, with one row per chain. NULL if the match is yet to start.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad in possession during the chain.
   \item \code{squad.code} A short code to represent the squad in possession during the chain.
   \item \code{period} The period of the possession chain.
   \item \code{start.id} The transaction ID of the start of the chain.
   \item \code{start.secs} The elapsed time within the period at the start of the chain.
   \item \code{start.name} The starting state of the chain.
   \item \code{start.code} A short code representing the starting state of the chain.
   \item \code{start.type} The start type of the chain.
   \item \code{start.type.code} A short code representing the starting type of the chain.
   \item \code{start.zone} The starting zone of the chain, relative to the squad in possession.
   \item \code{start.zone.physical} The starting zone of the chain, relative to the physical ground.
   \item \code{launch.id} A unique numerical identifier of the player who launched the chain.
   \item \code{launch.name} The full name of the player who launched the chain.
   \item \code{launch.display} The display name of the player who launched the chain, represented as first initial and surname.
   \item \code{end.id} The transaction ID of the end of the chain.
   \item \code{end.secs} The elapsed time within the period at the end of the chain.
   \item \code{end.name} The final state of the chain.
   \item \code{end.code} A short code representing the final state of the chain.
   \item \code{end.type} The end type of the chain.
   \item \code{end.type.code} A short code representing the end type of the chain.
   \item \code{end.zone} The final zone of the chain, relative to the squad in possession.
   \item \code{end.zone.physical} The final zone of the chain, relative to the physical ground.
   \item \code{score.id} The transaction ID of the scoring shot, if a score was generated during the chain.
   \item \code{score.result} The result scoring shot, if a score was generated during the chain.
   \item \code{score.points} The points generated by the scoring shot, if a score was generated during the chain.
   \item \code{score.player.id} A unique numerical identifier of the player who scored within the chain.
   \item \code{score.player.name} The full name of the player who scored within the chain.
   \item \code{score.player.display} The display name of the player who scored within the chain, represented as first initial and surname.
   \item \code{turnover.id} A unique numerical identifier of the player who turned the ball over at the end of the chain.
   \item \code{turnover.name} The full name of the player who turned the ball overa at the end of the chain.
   \item \code{turnover.display} The display name of the player who turned the ball overa at the end of the chain, represented as first initial and surname.
   \item \code{turnover.type} The type of turnover that was commited. Kick, handball or non-disposal.
   \item \code{turnover.state} The match state of the turnover that was commited. Clearing stoppage or General Play.
   \item \code{turnover.creation} The creation method of the turnover that was commited. Forced/Unforced Giveaway or Opposition Takeaway.
   \item \code{clearance.id} The transaction ID of the clearance that was won within the chain.
   \item \code{inside50.id} The transaction ID of the first inside 50 that was generated by the squad in possession within the chain.
   \item \code{rebound50.id} The transaction ID of the first rebound 50 that was generated by the squad in possession within the chain.
   \item \code{next.id} The transaction ID of the start of the next possession chain.
   \item \code{prevoius.id} The transaction ID of the start of the previous possession chain.
   \item \code{chain.metres.gained} Metres gained by a team within possession chains, combining measures towards attacking goal and away from defensive goal.
   \item \code{chain.metres.gained.reply} Chain Metres conceded to the opposition in the next chain after a turnover.
   \item \code{chain.metres.gained.net} Net Metres gained in a chain of possession after subtracting opposition Metres gained after a turnover. Equal to Chain Metres minus Chain Metres Reply.
}
}
\description{
Get a list of possession chains for a match.
}
\examples{
getChains(216085122)
}
