% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getChainsWhitelist}
\alias{getChainsWhitelist}
\title{getChainsWhitelist}
\format{
An object of class \code{character} of length 54.
}
\usage{
getChainsWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getChains function.
}
\keyword{internal}
