% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getCurrentSeason}
\alias{getCurrentSeason}
\title{Current Season}
\usage{
getCurrentSeason(leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A numerical value for season ID.
}
\description{
Get current season for a league and level.
}
\examples{
getCurrentSeason(1,1)
getCurrentSeason()
}
