% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getEntries}
\alias{getEntries}
\title{Match Entries}
\usage{
getEntries(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of inside 50 entries for a match, with one row per entry. NULL if the match is yet to start.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{entry.trx.id} The transaction ID of the entry, used for ordering chronologically.
   \item \code{entry.period} The period of the entry.
   \item \code{entry.period.secs} The elapsed time within the period of the entry.
   \item \code{entry.player.id} A unique numerical identifier of the player responsible for the entry.
   \item \code{entry.player.fullname} The fullname of the player responsible for the entry. 
   \item \code{entry.player.display.name} The display name of the player responsible for the entry.
   \item \code{entry.squad.id} A numerical identifier of the squad player responsible for the entry belongs to. 
   \item \code{entry.squad.name} The name of the squad the player responsible for the entry belongs to. 
   \item \code{entry.squad.code} A short code to represent the squad the player responsible for the entry belongs to. 
   \item \code{entry.type} The type of inside 50 entry.
   \item \code{entry.source.name} The game situation relative to the origin of a team's possession that results in an entry.
   \item \code{entry.source.type} Inside 50 entries where direct indicates the first chain after the specified entry.source.name 
   
   For centre bounce and midfield stoppage \code{entry.source.name}, direct indicates the entry was in the clearance chain \cr
   
   For repeat \code{entry.source.name}, direct indicates the entry was on the first turnover chain after an opposition rebound 50 \cr
   
   For transition \code{entry.source.name}, direct indicates the entry was in the same chain as the rebound 50
   \item \code{entry.kick.is.ground} For inside 50 entries that are kicks, was the entry from a ground kick (true/false).
   \item \code{entry.kick.foot} For inside 50 entries that are kicks, which foot did the kicking player use.
   \item \code{entry.kick.intent} For inside 50 entries that are kicks, the intent of the kicker. 
   \item \code{entry.kick.distance} For inside 50 entries that are kicks, the distance of the kick.
   \item \code{entry.kick.direction} For inside 50 entries that are kicks, the direction of the kick
   \item \code{entry.kick.launch.distance.to.goal} For inside 50 entries that are kicks, the straight line distance to goal from the launch point of the kick.
   \item \code{entry.kick.result.distance.to.goal} For inside 50 entries that are kicks, the straight line distance to goal from the result point of the kick.
   \item \code{entry.kick.pressure.name} For inside 50 entries that are kicks, the name of the type of pressure applied to the kicker.
   \item \code{entry.kick.pressure.code} For inside 50 entries that are kicks, a short code to describe the type of pressure applied to the kicker.
   \item \code{entry.kick.pressure.points} For inside 50 entries that are kicks, the number of points associated with the pressure type applied to the kicker.
   \item \code{entry.kick.launch.location.x} The x coordinate of the inside 50 launch location.
   \item \code{entry.kick.launch.location.y} The y coordinate of the inside 50 launch location.
   \item \code{entry.kick.launch.location.x.std} The x coordinate of the inside 50 launch location transformed to a standard ground dimension (160x141).
   \item \code{entry.kick.launch.location.y.std} The y coordinate of the inside 50 launch location transformed to a standard ground dimension (160x141).
   \item \code{entry.kick.launch.location.rotated.x} The x coordinate of the inside 50 launch location rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.launch.location.rotated.y} The y coordinate of the inside 50 launch location rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.launch.location.rotated.x.std} The x coordinate of the inside 50 launch location transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.launch.location.rotated.y.std} The x coordinate of the inside 50 launch location transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.result.location.x} The x coordinate of the inside 50 result location. 
   \item \code{entry.kick.result.location.y} The y coordinate of the inside 50 result location. 
   \item \code{entry.kick.result.location.x.std} The x coordinate of the inside 50 result location transformed to a standard ground dimension (160x141).
   \item \code{entry.kick.result.location.y.std} The y coordinate of the inside 50 result location transformed to a standard ground dimension (160x141).
   \item \code{entry.kick.result.location.rotated.x} The x coordinate of the inside 50 result location rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.result.location.rotated.y} The y coordinate of the inside 50 result location rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.result.location.rotated.x.std} The x coordinate of the inside 50 result location transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
   \item \code{entry.kick.result.location.rotated.y.std} The y coordinate of the inside 50 result location transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
}
}
\description{
Get a list of Inside 50 entries for a match.
}
\examples{
getEntries(216085122)
}
