% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getEntriesWhitelist}
\alias{getEntriesWhitelist}
\title{getEntriesWhitelist}
\format{
An object of class \code{character} of length 40.
}
\usage{
getEntriesWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getEntries function.
}
\keyword{internal}
