% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getFixtureExposedFields}
\alias{getFixtureExposedFields}
\title{getFixtureExposedFields}
\format{
An object of class \code{character} of length 47.
}
\usage{
getFixtureExposedFields
}
\description{
This vector contains all of the fields we're exposing in the getFixture() function
}
\keyword{internal}
