% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getFixtureWhitelist}
\alias{getFixtureWhitelist}
\title{getFixtureWhitelist}
\format{
An object of class \code{character} of length 54.
}
\usage{
getFixtureWhitelist
}
\description{
This vector contains all of the fields in the getFixture() function
}
\keyword{internal}
