% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getLadder}
\alias{getLadder}
\title{Ladder}
\usage{
getLadder(seasonId, roundNumber, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{roundNumber}{An integer indicating the round number you are interested in. Returns the ladder as it stood after that round had completed. Defaults to the current ladder, or end of Home & Away for past seasons.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.
\itemize{
   \item \code{season.id} A numerical identifier of a season.
   \item \code{round.number} The round number of the match. Continues to count up during finals.
   \item \code{position} The squads ladder position.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{ladder.points} The number of ladder points accumulated by the squad.
   \item \code{ladder.percentage} The squad's ladder percentage, calculated by points scored divided by points conceded.
   \item \code{played} The number of matches played by the squad.
   \item \code{win} The number of matches the squad has won.
   \item \code{win.pct} The percentage of completed matches the squad has won.
   \item \code{draw} The number of matches the squad has drawn.
   \item \code{loss} The number of matches the squad has lost.
   \item \code{total.for} The total number of points scored by the squad.
   \item \code{total.against} The total number of points conceded by the squad.
   \item \code{total.margin} The total points differential for the squad.
   \item \code{average.for} The average number of points scored by the squad per match.
   \item \code{average.against} The average number of points conceded by the squad per match.
   \item \code{average.margin} The average points differential for the squad per match.
}}
}
\value{
A data frame with the requested ladder, with one row per squad.
}
\description{
Get the ladder for a given season.
}
\examples{
getLadder(2022)
getLadder(2022,4)
getLadder()
}
