% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLadderExposedFields}
\alias{getLadderExposedFields}
\title{getLadderExposedFields}
\format{
An object of class \code{character} of length 19.
}
\usage{
getLadderExposedFields
}
\description{
This vector contains all the fields we're exposing in the getLadder function and their equivalent renames.
}
\keyword{internal}
