% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getLeaders}
\alias{getLeaders}
\title{Match Stat Leaders}
\usage{
getLeaders(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of the five leading players for selected metrics.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{code} The metric code (ALL_CAPS format).
   \item \code{name} The metric name (singular).
   \item \code{plural} The metric name (plural).
   \item \code{player.id} A unique numerical identifier of the player.
   \item \code{player.name} The full name of the player.
   \item \code{player.display} The display name of the player, represented as first initial and surname.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{value} The numeric metric value with no formatting applied.
   \item \code{display} The formatted metric value as a text string.
   \item \code{rank} Dense rank within metric, with ties presented as ties.method='min'.
   \item \code{order} Order within the metric with no ties. Ties in metric value are broken by ordering by player surname then firstname.
}
}
\description{
Get leading five players for selected stats for a match.
}
\examples{
getLeaders(216085122)
}
