% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeadersExposedFields}
\alias{getLeadersExposedFields}
\title{getLeadersExposedFields}
\format{
An object of class \code{character} of length 16.
}
\usage{
getLeadersExposedFields
}
\description{
This vector contains all the fields we're exposing in the getLeadersExposedFields function and their equivalent renames.
}
\keyword{internal}
