% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeadersWhitelist}
\alias{getLeadersWhitelist}
\title{getLeadersWhitelist}
\format{
An object of class \code{character} of length 16.
}
\usage{
getLeadersWhitelist
}
\description{
This vector contains all the possible columns in the getLeadersWhitelist() function (after the unnesting).
}
\keyword{internal}
