% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getLeague}
\alias{getLeague}
\title{League Details}
\usage{
getLeague(leagueId = 1, ...)
}
\arguments{
\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about a league and level.
\itemize{
   \item \code{league.id} A unique numerical identifier of the league.
   \item \code{league.name} The name of the league.
   \item \code{league.code} A short code representing the name of the league.
   \item \code{league.gender} A short code representing The gender of a league. "M" for Men's, "W" for Women's.
}
}
\description{
Get details about a league.
}
\examples{
getLeague(1)
getLeague()
}
