% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getLeagueId}
\alias{getLeagueId}
\title{League ID}
\usage{
getLeagueId(leagueCode = "AFL", ...)
}
\arguments{
\item{leagueCode}{A short code for a league. Defaults to AFL Men's Premiership.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A numerical value for league ID.
}
\description{
Get the ID of a known league.
}
\examples{
getLeagueId('AFL')
getLeagueId()
}
