% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeagueInfo_leagueLevelCombos}
\alias{getLeagueInfo_leagueLevelCombos}
\title{getLeagueInfo_leagueLevelCombos}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
getLeagueInfo_leagueLevelCombos
}
\description{
This vector contains all the possible combinations of league and level in order to gather all info in getLeagueInfo function.
}
\keyword{internal}
