% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeagueInfoExposedFields}
\alias{getLeagueInfoExposedFields}
\title{getLeagueInfoExposedFields}
\format{
An object of class \code{character} of length 10.
}
\usage{
getLeagueInfoExposedFields
}
\description{
This vector contains all the fields we're exposing in the getLeagueInfoExposedFields function and their equivalent renames.
}
\keyword{internal}
