% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getLeagueLevel}
\alias{getLeagueLevel}
\title{League Level Details}
\usage{
getLeagueLevel(leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about a league and level.
\itemize{
   \item \code{league.id} A unique numerical identifier of the league.
   \item \code{league.name} The name of the league.
   \item \code{league.code} A short code representing the name of the league.
   \item \code{league.gender} A short code representing The gender of a league. "M" for Men's, "W" for Women's.
   \item \code{level.id} A unique numerical identifier of the level.
   \item \code{level.name} The name of the level.
   \item \code{level.code} A short code representing the name of the level.
   \item \code{currentSeason.id} The current season ID.
}
}
\description{
Get details about a league and level.
}
\examples{
getLeagueLevel(1,1)
getLeagueLevel()
}
