% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getLeagueLevelId}
\alias{getLeagueLevelId}
\title{League Level ID}
\usage{
getLeagueLevelId(leagueCode = "AFL", levelCode = "SEN", ...)
}
\arguments{
\item{leagueCode}{A short code for a league. Defaults to AFL Men's Premiership.}

\item{levelCode}{A short code for a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a league ID and level ID.
\itemize{
   \item \code{league.id} A unique numerical identifier of the league.
   \item \code{level.id} A unique numerical identifier of the level.
}
}
\description{
Get the ID of a known level for a league.
}
\examples{
getLeagueLevelId('AFL','SEN')
getLeagueLevelId()
}
