% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeagueLevelsInfoWhitelist}
\alias{getLeagueLevelsInfoWhitelist}
\title{getLeagueLevelsInfoWhitelist}
\format{
An object of class \code{character} of length 12.
}
\usage{
getLeagueLevelsInfoWhitelist
}
\description{
This vector contains all of the fields possible from the getLeagueLevelsInfo() function
}
\keyword{internal}
