% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{getLeaguesLevelsInfo}
\alias{getLeaguesLevelsInfo}
\title{League/Level Metadata}
\usage{
getLeaguesLevelsInfo(leagueId, levelId, ...)
}
\arguments{
\item{leagueId}{A unique numerical identifier of a league. (ie. \code{leagueId = 3})}

\item{levelId}{A unique numerical identifier of a level. (ie. \code{level = 1})}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about a league and level.
\itemize{
   \item \code{current.season.id} The season ID of the current season for a particular league.
   \item \code{league.name} The name of the league.
   \item \code{league.code} A short code representing the name of the league.
   \item \code{league.id} A unique numerical identifier of the league.
   \item \code{level.name} The name of the level.
   \item \code{level.code} A short code representing the name of the level.
   \item \code{level.id} A unique numerical identifier of the level.
   \item \code{level.classification.id} A unique numerical identifier of the given levels' classification.
   \item \code{level.classification.code} A short code representing thegiven levels' classification \code{ie. 'AFLTIER1'}
   \item \code{league.gender} A short code representing The gender of a league. "M" for Men's, "W" for Women's.
   \item \code{competition.id} The competition ID for the particular league/level.
   \item \code{championship.id} The championship ID for the particular league/level.
}
}
\description{
Get details about all leagues and levels exposed in the Champion Data AFL API.
}
\examples{
getLeagueInfo(leagueId = 12)
}
