% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getLeaguesLevelsInfoExposedFields}
\alias{getLeaguesLevelsInfoExposedFields}
\title{getSquadsExposedFields_all}
\format{
An object of class \code{character} of length 12.
}
\usage{
getLeaguesLevelsInfoExposedFields
}
\description{
This vector contains all of the fields and renames in the getLeaguesLevelsInfo() function
}
\keyword{internal}
