% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ref.R
\name{getMatch}
\alias{getMatch}
\title{Match Details}
\usage{
getMatch(matchId, ...)
}
\arguments{
\item{matchId}{An unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about a match.
\itemize{
   \item \code{competition.name} The full name of the competition.
   \item \code{season.id} A numerical identifier of a season.
   \item \code{type.id} A numerical indicator of the match type.
   \item \code{type.name} A description of the match type.
   \item \code{round.number} The round number of the match. Continues to count up during finals.
   \item \code{round.phaseNumber} The round number of the match within the phase. Starts again from one for finals.
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{match.name} The name of the match. Concatenates home and away squad names.
   \item \code{match.code} A short code for the name of the match. Concatenates home and away squad codes.
   \item \code{match.start} The start time of a match in UTC time.
   \item \code{match.date} The calendar date of a match in local time.
   \item \code{match.time} The local start time of a match.
   \item \code{home.id} A unique numerical identifier of the home squad.
   \item \code{home.name} The name of the home squad.
   \item \code{home.code} A short code to represent the home squad.
   \item \code{away.id} A unique numerical identifier of the away squad.
   \item \code{away.name} The name of the away squad.
   \item \code{away.code} A short code to represent the away squad.
   \item \code{venue.id} A unique numerical identifier of the match venue.
   \item \code{venue.name} The name of the match venue.
   \item \code{venue.code} A short code to represent the match venue.
   \item \code{venue.timezone} The timezone of the venue.
   \item \code{status.id} A numerical identifier of the match's current status.
   \item \code{status.name} The match's current status.
   \item \code{status.type.id} A numerical identifier of the match status type.
   \item \code{status.type.name} The match's current status type.
   \item \code{home.goals} The number of goals kicked by the home squad.
   \item \code{home.behinds} The number of behinds kicked by the home squad.
   \item \code{home.points} The number of points scored by the home squad.
   \item \code{away.goals} The number of goals kicked by the away squad.
   \item \code{away.behinds} The number of behinds kicked by the away squad.
   \item \code{away.points} The number of points scored by the away squad.
   \item \code{winner.id} The squad ID of the winning team.
   \item \code{result} A text description of the match result.
   \item \code{result.code} A short text description of the match result.
}
}
\description{
Get details about a match.
}
\examples{
getMatch(216085122)
}
