% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatch_PeriodsExposedFields}
\alias{getMatch_PeriodsExposedFields}
\title{getMatch_PeriodsExposedFields}
\format{
An object of class \code{character} of length 5.
}
\usage{
getMatch_PeriodsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getMatch function when periods = TRUE and their equivalent renames.
}
\keyword{internal}
