% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatch_PeriodsWhitelist}
\alias{getMatch_PeriodsWhitelist}
\title{getMatch_PeriodsWhitelist}
\format{
An object of class \code{character} of length 4.
}
\usage{
getMatch_PeriodsWhitelist
}
\description{
This vector contains all the possible fields in the getMatch function when periods = TRUE
}
\keyword{internal}
