% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchExposedFields}
\alias{getMatchExposedFields}
\title{getMatchExposedFields}
\format{
An object of class \code{character} of length 51.
}
\usage{
getMatchExposedFields
}
\description{
This vector contains all the fields we're exposing in the getMatch function and their equivalent renames.
}
\keyword{internal}
