% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ref.R
\name{getMatchPeriods}
\alias{getMatchPeriods}
\title{Match Periods}
\usage{
getMatchPeriods(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about each match period. Only contains periods that have started.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{period} The period of the match.
   \item \code{secs} The elapsed time in seconds for live periods, or the period length for completed periods.
   \item \code{utcPeriodStart} The start time of the period in UTC time.
   \item \code{utcPeriodEnd} The end time of the period in UTC time.
}
}
\description{
Get details about the periods of a match.
}
\examples{
getMatchPeriods(216085122)
}
