% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersons_homeWhitelist}
\alias{getMatchPersons_homeWhitelist}
\title{getMatchPersons_homeWhitelist}
\format{
An object of class \code{character} of length 18.
}
\usage{
getMatchPersons_homeWhitelist
}
\description{
This vector contains all the possible columns in the getMatchPersons() function for when the *HOME* team is split out.
}
\keyword{internal}
