% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersons_officialsDF}
\alias{getMatchPersons_officialsDF}
\title{getMatchPersons_officialsDF}
\format{
An object of class \code{data.frame} with 0 rows and 8 columns.
}
\usage{
getMatchPersons_officialsDF
}
\description{
This is the empty dataframe returned in the response when officials param of getMatchPersons() is set to TRUE, but there are no officials in the response (ie. for AFLW matches where officials arent recorded).
}
\keyword{internal}
