% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersons_officialsExposedFields}
\alias{getMatchPersons_officialsExposedFields}
\title{getMatchPersons_officialsExposedFields}
\format{
An object of class \code{character} of length 8.
}
\usage{
getMatchPersons_officialsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getMatchPersons function and their equivalent renames for the officials portion of the response (officials param == TRUE).
}
\keyword{internal}
