% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersons_officialsWhitelist}
\alias{getMatchPersons_officialsWhitelist}
\title{getMatchPersons_officialsWhitelist}
\format{
An object of class \code{character} of length 8.
}
\usage{
getMatchPersons_officialsWhitelist
}
\description{
This vector contains all the possible columns in the getMatchPersons() function for the officials portion of the response (officials param == TRUE).
}
\keyword{internal}
