% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersons_Whitelist}
\alias{getMatchPersons_Whitelist}
\title{getMatchPersons_Whitelist}
\format{
An object of class \code{character} of length 18.
}
\usage{
getMatchPersons_Whitelist
}
\description{
This vector contains all the possible columns in the getMatchPersons() function
}
\keyword{internal}
