% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchPersonsExposedFields}
\alias{getMatchPersonsExposedFields}
\title{getMatchPersonsExposedFields}
\format{
An object of class \code{character} of length 20.
}
\usage{
getMatchPersonsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getMatchPersonsExposedFields function and their equivalent renames.
}
\keyword{internal}
