% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getMatchTransactions}
\alias{getMatchTransactions}
\title{Match Transactions}
\usage{
getMatchTransactions(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of transactions for a match. 
\itemize{
   \item \code{match.id} A unique numerical identifier for a match.
   \item \code{trx.id} A unique numerical identifier for a given transaction, used for ordering chronologically.
   \item \code{stat.code} The statistic code for a given transaction.
   \item \code{stat.desc} A short description of the statistic code for a transaction.  
   \item \code{period} The period of the transaction.
   \item \code{period.secs} The elapsed time within the period of a transaction. 
   \item \code{score.home} The score of the home squad.
   \item \code{score.away} The score of the away squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{squad.id} A unique numerical identifier for the squad.
   \item \code{person.fullname} The fullname of the player assigned to the transaction (if applicable).
   \item \code{person.displayName} The display name of the player assigned to the transaction (if applicable).
   \item \code{person.id} A unique numerical identifier for the player assigned to the transaction (if applicable).
   \item \code{pressure.squad.name} The name of the squad the pressure information is assigned to.
   \item \code{pressure.squad.id} A unique numerical identifier for the squad the pressure information is assigned to. 
   \item \code{pressure.squad.code} A short code to represent the squad the pressure information is assigned to.
   \item \code{pressure.name} The name of the type of pressure applied (ie. Set Position). 
   \item \code{pressure.code} A short code to represent the type of pressure applied.
   \item \code{pressure.points} The points value of the type of pressure applied.
   \item \code{pressure.fullname1} The fullname of the first pressure player.
   \item \code{pressure.displayname1} The display name of the first pressure player.
   \item \code{pressure.personid1} A unique numerical identifier for the first pressure player.
   \item \code{pressure.role1} The role number of the first pressure player (1).
   \item \code{pressure.fullname2} The fullname of the second pressure player.
   \item \code{pressure.displayname2} The display name of the second pressure player. 
   \item \code{pressure.personid2} A unique numerical identifier for the second pressure player.
   \item \code{pressure.role2} The role number of the second pressure player (2).
   \item \code{zone.physical} The zone in which the transaction takes place, relative to the physical ground.
   \item \code{zone.logical} The zone in which the transaction takes place, relative to the squad in possession.
   \item \code{location.x} The x coordinate of the transaction.
   \item \code{location.y} the y coordinate of the transaction.
   \item \code{location.x.std} The x coordinate of the transaction transformed to a standard ground dimension (160x141).
   \item \code{location.y.std} The y coordinate of the transaction transformed to a standard ground dimension (160x141).
   \item \code{location.rotated.x} The x coordinate of the transaction rotated to give all squads attacking left-to-right.
   \item \code{location.rotated.y} The y coordinate of the transaction rotated to give all squads attacking left-to-right.
   \item \code{location.rotated.x.std} The x coordinate of the transaction transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
   \item \code{location.rotated.y.std} The y coordinate of the transaction transformed to a standard ground dimension (160x141) & rotated to give all squads attacking left-to-right.
   \item \code{kicking.foot} On kicking transactions, the foot of the kicking player.
   \item \code{kicking.intent} On kicking transactions, the intent of the kicking player.
   \item \code{kicking.distance} On kicking transactions, a description of the kicks distance.
   \item \code{kicking.direction} On kicking transactions, a description of the kicks direction.
   \item \code{inside50.direction} On inside 50 transactions, a description of the kicks direction.
   \item \code{inside50.intent} On inside 50 transactions, the intent of the kicking player.
   \item \code{shot.angle} A text description of the angle of the shot.
   \item \code{shot.distance} A text description of the distance of the shot.
   \item \code{shot.result} The result of the shot on goal (Does not include Goals).
   \item \code{shot.source} The source from which the shot on goal resulted (ie. General Play).
   \item \code{shot.type} The type of shot attempted (ie. General Play Snap).
   \item \code{freekick.context} On free kick transactions, the broad context from which the free kick was awarded.
   \item \code{freekick.reason} On free kick transactions, the specific reason the free kick was awarded.
   \item \code{kickin.direction} On kick in transactions, a text description of the direction taken from the kick in.
}
}
\description{
Get the list of transactions for a match. Returns observed & not yet observed metrics.
}
\examples{
getMatchTransactions(216085122)
}
