% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchTransactionsExposedFields}
\alias{getMatchTransactionsExposedFields}
\title{getMatchTransactionsExposedFields}
\format{
An object of class \code{character} of length 52.
}
\usage{
getMatchTransactionsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getMatchTransactions function and their equivalent renames.
}
\keyword{internal}
