% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getMatchTransactionsPressureCols}
\alias{getMatchTransactionsPressureCols}
\title{getMatchTransactionsPressureCols}
\format{
An object of class \code{character} of length 8.
}
\usage{
getMatchTransactionsPressureCols
}
\description{
This vector contains the column names of the pressure columns split out in the getMatchTransactions function.
}
\keyword{internal}
