% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchTransactionsWhitelist}
\alias{getMatchTransactionsWhitelist}
\title{getMatchTransactionsWhitelist}
\format{
An object of class \code{character} of length 52.
}
\usage{
getMatchTransactionsWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getMatchTransactions function.
}
\keyword{internal}
