% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMatchWhitelist}
\alias{getMatchWhitelist}
\title{getMatchWhitelist}
\format{
An object of class \code{character} of length 48.
}
\usage{
getMatchWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getMatch function.
}
\keyword{internal}
