% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMetricsExposedFields}
\alias{getMetricsExposedFields}
\title{getMetricsExposedFields}
\format{
An object of class \code{character} of length 16.
}
\usage{
getMetricsExposedFields
}
\description{
This vector contains all of the fields we're exposing in the getMetrics() function
}
\keyword{internal}
