% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getMetricsWhitelist}
\alias{getMetricsWhitelist}
\title{getMetricsWhitelist}
\format{
An object of class \code{character} of length 15.
}
\usage{
getMetricsWhitelist
}
\description{
This vector contains all of the fields in the getMetrics() function
}
\keyword{internal}
