% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPeriodScoresExposedFields}
\alias{getPeriodScoresExposedFields}
\title{getPeriodScoresExposedFields}
\format{
An object of class \code{character} of length 22.
}
\usage{
getPeriodScoresExposedFields
}
\description{
This vector contains all of the fields we're exposing when cumulative = FALSE in the getPeriodScore() function
}
\keyword{internal}
