% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPeriodScoresExposedFields_cumulative}
\alias{getPeriodScoresExposedFields_cumulative}
\title{getPeriodScoreExposedFields_cumulative}
\format{
An object of class \code{character} of length 22.
}
\usage{
getPeriodScoresExposedFields_cumulative
}
\description{
This vector contains all of the fields we're exposing when cumulative = TRUE in the getPeriodScore() function
}
\keyword{internal}
