% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPeriodScoresWhitelist}
\alias{getPeriodScoresWhitelist}
\title{getPeriodScoresWhitelist}
\format{
An object of class \code{character} of length 44.
}
\usage{
getPeriodScoresWhitelist
}
\description{
This vector contains all the possible columns in the info portion of the getPeriodScores() function.
}
\keyword{internal}
