% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getPlayerStats}
\alias{getPlayerStats}
\title{Match Player Stats}
\usage{
getPlayerStats(
  matchId,
  period = NULL,
  zone = NULL,
  metric = NULL,
  team = NULL,
  lastXseconds = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{period}{A numerical indicator of a period to filter results within a match. Accepts integer values.}

\item{zone}{A text indicator of a zone on the field to filter results within a match. Accepts string values: \code{"D50","DM","AM","F50","CB"}.}

\item{metric}{A text string of specific metric code(s) to be returned. This will result in the endpoint only returning these specific metrics. Note this endpoint is case sensitive and only works with metric codes (ie. \code{c("TACKLE", "GOAL")}}

\item{team}{A (case-sensitive) text string of the team to return metrics for. Either \code{"home"} or \code{"away"}. Not passing anything to this param will return both teams.}

\item{lastXseconds}{An integer that limits statistics to counting events that occurred in the last X number of seconds. (ie. \code{lastXSeconds = 300} will return the last 5 minutes)}

\item{from}{Limits statistics to counting events that occurred on or after this number of seconds (ie. \code{from = 300} will return statistics that occured on or after the 300th second into the match)}

\item{to}{Limits statistics to counting events that occurred on or before this number of seconds (ie. \code{to = 300} will return statistics that occured up to (and including) the 300th second into the match)}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of metrics for a match for each player.
\itemize{
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{player.id} A unique numerical identifier of the player.
   \item \code{player.name} The full name of the player.
   \item \code{player.display} The display name of the player, represented as first initial and surname.
   \item \code{code} The metric code (ALL_CAPS format).
   \item \code{name} The metric name (singular).
   \item \code{id} The unique numerical identifier for the given statistic code.
   \item \code{plural} The metric name (plural).
   \item \code{value} The numeric metric value with no formatting applied.
   \item \code{display} The formatted metric value as a text string.
}
}
\description{
Get player stats for a match. Only returns observed metrics.
}
\examples{
getPlayerStats(216085122)
getPlayerStats(216085122,period=1,zone='D50',from = 0,to = 300)
}
