% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getPlayerStatsMissingFields}
\alias{getPlayerStatsMissingFields}
\title{getPlayerStatsMissingFields}
\format{
An object of class \code{character} of length 5.
}
\usage{
getPlayerStatsMissingFields
}
\description{
This vector contains all the fields that are not present when there is no data to unnest following the tidyr::unnest("squads.players_statistics"...) call in the code. This vector adds the fields listed as columns so the function can return a consistent output.
}
\keyword{internal}
