% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPlayerStatsPOST_info_Whitelist}
\alias{getPlayerStatsPOST_info_Whitelist}
\title{getPlayerStatsPOST_info_Whitelist}
\format{
An object of class \code{character} of length 5.
}
\usage{
getPlayerStatsPOST_info_Whitelist
}
\description{
This vector contains all the possible columns in the info portion of the getPlayerStatsPOST() function.
}
\keyword{internal}
