% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPlayerStatsPOSTExposedFields}
\alias{getPlayerStatsPOSTExposedFields}
\title{getPlayerStatsPOSTExposedFields}
\format{
An object of class \code{character} of length 14.
}
\usage{
getPlayerStatsPOSTExposedFields
}
\description{
This vector contains all the fields we're exposing in the getPlayerStatsPOST function (not including info) and their equivalent renames.
}
\keyword{internal}
