% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPlayerSummaryFileExposedFields}
\alias{getPlayerSummaryFileExposedFields}
\title{getPlayerSummaryFileExposedFields}
\format{
An object of class \code{character} of length 240.
}
\usage{
getPlayerSummaryFileExposedFields
}
\description{
This vector contains all of the fields we're exposing in the getPlayerSummaryFile() function
}
\keyword{internal}
