% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getRotations_HomeAwayStintsExposedFields}
\alias{getRotations_HomeAwayStintsExposedFields}
\title{getRotations_HomeAwayStintsExposedFields}
\format{
An object of class \code{character} of length 12.
}
\usage{
getRotations_HomeAwayStintsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getRotations function when currentStints = TRUE and their equivalent renames.
}
\keyword{internal}
