% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getRotationsExposedFields}
\alias{getRotationsExposedFields}
\title{getRotationsExposedFields}
\format{
An object of class \code{character} of length 15.
}
\usage{
getRotationsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getRotations function and their equivalent renames.
}
\keyword{internal}
