% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getRotationsWhitelist}
\alias{getRotationsWhitelist}
\title{getRotationsWhitelist}
\format{
An object of class \code{character} of length 15.
}
\usage{
getRotationsWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getRotations function.
}
\keyword{internal}
