% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getRuckContests}
\alias{getRuckContests}
\title{Match Ruck Contest Summary}
\usage{
getRuckContests(matchId, stoppageType = NULL, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{stoppageType}{Filter statistics to a specific type(s) of ruck contest, any one of: \cr
\code{'centreBounce', 'ballUp', 'throwIn'}\cr

Can accept multiple arguments in a character vector\cr(ie. \code{stoppageType = c('ballUp', 'throwIn')})}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a summary of ruck contest statistics for a given match.
\itemize{
 \item \code{match.id}: The unique identifier for the match.
 \item \code{total.contests}: The total number of ruck contests between the home & away ruck combination.
 \item \code{home.ruck.squad.code}: Squad code for the home team's ruckman.
 \item \code{home.ruck.squad.id}: Squad ID for the home team's ruckman.
 \item \code{home.ruck.fullname}: Fullname of the home team's ruckman.
 \item \code{home.ruck.display.name}: Display name of the home team's ruckman.
 \item \code{home.ruck.person.id}: Person ID of the home team's ruckman.
 \item \code{away.ruck.squad.code}: Squad code for the away team's ruckman.
 \item \code{away.ruck.squad.id}: Squad ID for the away team's ruckman.
 \item \code{away.ruck.fullname}: Fullname of the away team's ruckman.
 \item \code{away.ruck.display.name}: Display name of the away team's ruckman.
 \item \code{away.ruck.person.id}: Person ID of the away team's ruckman.
 \item \code{home.ruck.hitouts}: Total number of hitouts by the home team's ruckman.
 \item \code{home.ruck.hitouts.advantage}: Total number of hitouts to advantage by the home team's ruckman.
 \item \code{home.ruck.hitouts.sharked}: Total number of hitouts sharked from the home team's ruckman.
 \item \code{home.ruck.first.possessions}: Total number of first possessions gathered by the home team while the given home ruck was in the ruck contest.  
 \item \code{home.ruck.clearances}: Total number of clearances gathered by the home team while the given home ruck was in the ruck contest.  
 \item \code{home.ruck.goals}: Total number of goals gathered by the home team while the given home ruck was in the ruck contest (in home team clearance chains).
 \item \code{home.ruck.behinds}: Total number of behinds gathered by the home team while the given home ruck was in the ruck contest (in home team clearance chains).
 \item \code{home.ruck.points}: Total number of points gathered by the home team while the given home ruck was in the ruck contest (in home team clearance chains).
 \item \code{away.ruck.hitouts}: Total number of hitouts by the away team's ruckman.
 \item \code{away.ruck.hitouts.advantage}: Total number of hitouts to advantage by the away team's ruckman.
 \item \code{away.ruck.hitouts.sharked}: Total number of hitouts sharked from the away team's ruckman.
 \item \code{away.ruck.first.possessions}: Total number of first possessions gathered by the away team while the given away ruck was in the ruck contest.  
 \item \code{away.ruck.clearances}: Total number of clearances gathered by the away team while the given away ruck was in the ruck contest.
 \item \code{away.ruck.goals}: Total number of goals gathered by the away team while the given away ruck was in the ruck contest (in away team clearance chains).
 \item \code{away.ruck.behinds}: Total number of behinds gathered by the away team while the given away ruck was in the ruck contest (in away team clearance chains).
 \item \code{away.ruck.points}: Total number of points gathered by the away team while the given away ruck was in the ruck contest (in away team clearance chains).
}
}
\description{
Get a summary of ruck contest statistics for a given match.
}
\examples{
getRuckContests(120390401)
getRuckContests(120390401, stoppageType=c('ballUp', 'throwIn'))
}
