% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getRuckContestsExposedFields}
\alias{getRuckContestsExposedFields}
\title{getRuckContestsExposedFields}
\format{
An object of class \code{character} of length 28.
}
\usage{
getRuckContestsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getRuckContests function and their equivalent renames.
}
\keyword{internal}
