% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getSeason}
\alias{getSeason}
\title{Season Details}
\usage{
getSeason(seasonId, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about the given season.
\itemize{
   \item \code{season.id} A numerical identifier of a season.
   \item \code{competition.id} A unique numerical identifier of a competition, combining League, Level and Season.
   \item \code{competition.code} A short code describing the competition.
   \item \code{competition.name} The full name of the competition.
   \item \code{startDate} The date of the first match of the season 'YYYY-MM-DD'.
   \item \code{startYear} The year of the first match of the season.
   \item \code{endDate} The date of the final scheduled game of the season 'YYYY-MM-DD'.
   \item \code{endYear} The year of the final scheduled game of the season.
   \item \code{complete} A count of the complete matches in the season.
   \item \code{incomplete} A count of incomplete matches remaining in the season.
   \item \code{nextRound} The round number of the next round to start.
   \item \code{nextRoundStart} The datetime of the first match of the next round to start, in UTC time.
   \item \code{lastRound} The round number of the last round to finish.
   \item \code{lastRoundEnd} The datetime of the end of the last match of the last round to finish, in UTC time.
   \item \code{firstMatchStart} The datetime of the first match of the season.
}
}
\description{
Get details about a given season.
}
\examples{
getSeason(2022)
getSeason()
}
