% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getSeasonExposedFields}
\alias{getSeasonExposedFields}
\title{getShotsExposedFields}
\format{
An object of class \code{character} of length 15.
}
\usage{
getSeasonExposedFields
}
\description{
This vector contains all the fields we're exposing in the getSeason function and their equivalent renames.
}
\keyword{internal}
