% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getSeasons}
\alias{getSeasons}
\title{Seasons Info}
\usage{
getSeasons(leagueId = 1, levelId = 1, seasonId, ...)
}
\arguments{
\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors (1).}

\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of metrics for a match for each player.
\itemize{
  \item \code{competition.name} The name of the competition.
  \item \code{competition.code} A short code describing the competition.
  \item \code{competition.id} A unique numerical identifier of the competition.
  \item \code{season.id} A numerical identifier of a season.
  \item \code{season.start.date} The start date of the season.
  \item \code{season.end.date} The end date of the season.
  \item \code{league.id} A unique numerical identifier of the league.
  \item \code{league.name} The name of the league.
  \item \code{league.code} A short code representing the name of the league. 
  \item \code{league.gender} A short code representing The gender of a league. "M" for Men's, "W" for Women's.
  \item \code{matches.complete} The number of matches in the season that have been completed.
  \item \code{matches.incomplete} The number of matches as-yet incomplete for the season.
  \item \code{first.match.start} The date-time of the first match of the season.
  \item \code{last.completed.round.number} The round number of the last completed round of the next scheduled round of the season.
  \item \code{last.completed.round.start} The start date-time of the last completed round of the season.
  \item \code{last.completed.round.end} The end date-time of the last completed round of the season.
  \item \code{next.scheduled.round.number} The round number of the next scheduled round of the season.
  \item \code{next.scheduled.round.start} The start date-time of the next scheduled round of the season.
  \item \code{next.scheduled.round.end} The end date-time of the next scheduled round of the season.
}
}
\description{
Get information about all available seasons for a given league/level.
}
\examples{
getSeasons()
getSeasons(leagueId = 12)
}
