% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSeasonsExposedFields}
\alias{getSeasonsExposedFields}
\title{getSeasonsExposedFields}
\format{
An object of class \code{character} of length 19.
}
\usage{
getSeasonsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getSeasons function and their equivalent renames.
}
\keyword{internal}
