% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSeasonWhitelist}
\alias{getSeasonWhitelist}
\title{getSeasonWhitelist}
\format{
An object of class \code{character} of length 27.
}
\usage{
getSeasonWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getSeason function.
}
\keyword{internal}
