% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getShots}
\alias{getShots}
\title{Match Shots}
\usage{
getShots(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of shots at goal for a match, with one row per shot.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{id} The transaction ID of the shot, used for ordering chronologically.
   \item \code{period} The period of the shot.
   \item \code{secs} The elapsed time within the period of the shot.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{player.id} A unique numerical identifier of the player taking the shot.
   \item \code{player.name} The full name of the player taking the shot.
   \item \code{player.display} The display name of the player taking the shot, represented as first initial and surname.
   \item \code{origin} How the player was able to generate a shot at goal.
   \item \code{type} The type of shot attempted.
   \item \code{angle} A text description of the angle of the shot.
   \item \code{distance} A text description of the distance of the shot.
   \item \code{x} Distance in metres from the centre of the ground (measuring goal to goal), with the goals being at positive (venue_length / 2).
   \item \code{y} Distance in metres from the centre of the ground (measuring boundary to boundary), with the goals being at zero and the boundary line at halfway being at (venue_widht / 2). Positive y is bottom of screen when running to the right.
   \item \code{x.std} The x coordinate of the shot transformed to a standard ground dimension (160x141).
   \item \code{y.std} The y coordinate of the shot transformed to a standard ground dimension (160x141).
   \item \code{result} The result of the shot. G / B / R / M.
   \item \code{points} The number of points obtained from the shot.
}
}
\description{
Get a list of shots at goal for a match. Also includes rushed behinds.
}
\examples{
getShots(216085122)
}
