% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getShotsExposedFields}
\alias{getShotsExposedFields}
\title{getShotsExposedFields}
\format{
An object of class \code{character} of length 22.
}
\usage{
getShotsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getShots function and their equivalent renames.
}
\keyword{internal}
