% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getShotsWhitelist}
\alias{getShotsWhitelist}
\title{getShotsWhitelist}
\format{
An object of class \code{character} of length 29.
}
\usage{
getShotsWhitelist
}
\description{
This vector contains all the possible columns that can be returned by hitting the endpoint within the getShots function.
}
\keyword{internal}
