% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getSquad}
\alias{getSquad}
\title{Squad Details}
\usage{
getSquad(squadId, seasonId, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{squadId}{A unique numerical identifier of a squad.}

\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with details about a squads.
\itemize{
   \item \code{season.id} A numerical identifier of a season.
   \item \code{id} A unique numerical identifier of the squad.
   \item \code{name} The name of the squad.
   \item \code{code} A short code to represent the squad.
   \item \code{state.id} A numerical identifier of a country's state.
   \item \code{state.name} The squad's home state.
   \item \code{state.code} A short code representing the squad's home state.
}
}
\description{
Get details about a squad in a season.
}
\examples{
getSquad(10,2022)
getSquad(20)
}
