% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getSquadLists}
\alias{getSquadLists}
\title{Squad Person Lists}
\usage{
getSquadLists(seasonId, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of persons for all squads, with one row per person.
\itemize{
   \item \code{season.id} A numerical identifier of a season.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{id} A unique numerical identifier of a person.
   \item \code{firstname} The person's first name.
   \item \code{surname} The person's surname.
   \item \code{name} The person's full name.
   \item \code{display} The person's display name, represented as first initial and surname.
   \item \code{position} The person's primary position within a season. The position where a player has spent most time.
   \item \code{DOB} The person's date of birth (YYYY-MM-DD).
   \item \code{age.season} The person's age in years at the start of the first game of the season, to one decimal place.
   \item \code{age.year} The person's age at December 31 of the year in which the season ends, rounded down to a whole year.
   \item \code{height} The person's height in centimetres.
   \item \code{weight} The person's weight in kilograms.
}
}
\description{
Get a list of persons involved in all squads in a season. Currently limited to the playing list.
}
\examples{
getSquadPersons(10,2022)
getSquadPersons('Carlton')
}
