% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadPersonsExposedFields}
\alias{getSquadPersonsExposedFields}
\title{getSquadPersonsExposedFields}
\format{
An object of class \code{character} of length 21.
}
\usage{
getSquadPersonsExposedFields
}
\description{
This vector contains all of the fields we're exposing in the getSquadPersons() function
}
\keyword{internal}
