% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadsExposedFields_all}
\alias{getSquadsExposedFields_all}
\title{getSquadsExposedFields_all}
\format{
An object of class \code{character} of length 4.
}
\usage{
getSquadsExposedFields_all
}
\description{
This vector contains all of the fields we're exposing in the getSquads() function WHEN a squadId IS NOT passed in
}
\keyword{internal}
