% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadsExposedFields_squad}
\alias{getSquadsExposedFields_squad}
\title{getSquadsExposedFields_squad}
\format{
An object of class \code{character} of length 7.
}
\usage{
getSquadsExposedFields_squad
}
\description{
This vector contains all of the fields we're exposing in the getSquads() function WHEN a squadId is passed in
}
\keyword{internal}
