% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getSquadStats}
\alias{getSquadStats}
\title{Match Squad Stats}
\usage{
getSquadStats(matchId, period, zone, context, metric, team, lastXseconds, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{period}{A numerical indicator of a period to filter results within a match. Accepts integer values.}

\item{zone}{A text indicator of a zone on the field to filter results within a match. Accepts string values "D50","DM","AM","F50".}

\item{context}{A text indicator of the context for squad statistics. Accepts "For" "Against" or "Diff".}

\item{metric}{A text string of specific metric code(s) to be returned. This will result in the endpoint only returning these specific metrics. Note this endpoint is case sensitive and only works with metric codes (ie. \code{c("TACKLE", "GOAL")}}

\item{team}{A (case-sensitive) text string of the team to return metrics for. Either \code{"home"} or \code{"away"}. Not passing anything to this param will return both teams.}

\item{lastXseconds}{An integer that limits statistics to counting events that occurred in the last X number of seconds. (ie \code{lastXSeconds = 300} will return the last 5 minutes.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of metrics for a match for each squad.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{code} The metric code (ALL_CAPS format).
   \item \code{name} The metric name (singular).
   \item \code{plural} The metric name (plural).
   \item \code{value} The numeric metric value with no formatting applied.
   \item \code{display} The formatted metric value as a text string.
}
}
\description{
Get squad stats for a match. Only returns observed metrics.
}
\examples{
getSquadStats(216085122)
getSquadStats(216085122,period=1,zone='D50')
}
