% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadStatsExposedFields}
\alias{getSquadStatsExposedFields}
\title{getSquadStatsExposedFields}
\format{
An object of class \code{character} of length 10.
}
\usage{
getSquadStatsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getSquadStats function and their equivalent renames.
}
\keyword{internal}
