% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadStatsMissingFields}
\alias{getSquadStatsMissingFields}
\title{getSquadStatsMissingFields}
\format{
An object of class \code{character} of length 6.
}
\usage{
getSquadStatsMissingFields
}
\description{
This vector contains all the fields that are not present when there is no data to unnest following the tidyr::unnest("squads.statistics"...) call in the code. This vector adds the fields listed as columns so the function can return a consistent output.
}
\keyword{internal}
