% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadStatsPOST_info_ExposedFields}
\alias{getSquadStatsPOST_info_ExposedFields}
\title{getSquadStatsPOST_info_ExposedFields}
\format{
An object of class \code{character} of length 7.
}
\usage{
getSquadStatsPOST_info_ExposedFields
}
\description{
This vector contains all the fields we're exposing in the getPlayerStatsPOST function INFO and their equivalent renames.
}
\keyword{internal}
