% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadStatsPOST_info_Whitelist}
\alias{getSquadStatsPOST_info_Whitelist}
\title{getSquadStatsPOST_info_Whitelist}
\format{
An object of class \code{character} of length 6.
}
\usage{
getSquadStatsPOST_info_Whitelist
}
\description{
This vector contains all the possible columns in the info portion of the getSquadStatsPOST() function.
}
\keyword{internal}
