% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getSquadSummaryFile}
\alias{getSquadSummaryFile}
\title{AFL Squad Summary File}
\usage{
getSquadSummaryFile(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame in the same format and structure as the AFL Squad Summary File formerly available via the support site.
\itemize{
   \item \code{MATCH_ID} A unique numerical identifier of a match.
   \item \code{SEASON_ID} A numerical identifier of a season.
   \item \code{GROUP_ROUND_NO} The round number of the match. Continues to count up during finals.
   \item \code{VENUE_NAME} The name of the match venue.
   \item \code{SQUAD_NAME} The name of the squad for the given row.
   \item \code{OPP_SQUAD_NAME} The name of the opposition squad for the given row.
   \item \code{SQUAD_MARGIN} The final margin of the match in relation to the team listed under SQUAD_NAME.
   \item \code{MARGIN} The margin at the end of the given period in relation to the team listed under SQUAD_NAME.
   \item \code{ZONE_LOGICAL_AFL} The zone in which the values within metrics that follow are belonging to, relative to the squad listed under SQUAD_NAME.
}
}
\description{
Get the AFL Squad Summary File formerly available via the support site.
}
\examples{
getSquadSummaryFile(matchId = 120390401)
}
