% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadSummaryFileExposedFields}
\alias{getSquadSummaryFileExposedFields}
\title{getSquadSummaryFileExposedFields}
\format{
An object of class \code{character} of length 314.
}
\usage{
getSquadSummaryFileExposedFields
}
\description{
This vector contains all of the fields we're exposing in the getSquadSummaryFile() function
}
\keyword{internal}
