% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getSquadsWhitelist}
\alias{getSquadsWhitelist}
\title{getSquadsWhitelist}
\format{
An object of class \code{character} of length 7.
}
\usage{
getSquadsWhitelist
}
\description{
This vector contains all of the fields in the getSquads() function
}
\keyword{internal}
