% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getStatisticFlow}
\alias{getStatisticFlow}
\title{Statistic Flow}
\usage{
getStatisticFlow(
  matchId,
  metric,
  period = NULL,
  zone = NULL,
  team = NULL,
  limit = NULL,
  orderBy = NULL,
  ...
)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{metric}{A text string of specific metric code(s) to be returned. This will result in the endpoint only returning these specific metrics. Note this endpoint is case sensitive and only works with metric codes (ie. \code{c("TACKLE", "GOAL")}}

\item{period}{A numerical indicator of a period to filter results within a match. Accepts integer values.}

\item{zone}{A text indicator of a zone on the field to filter results within a match. Accepts string values: \code{"D50","DM","AM","F50","CB"}.}

\item{team}{A (case-sensitive) text string of the team to return metrics for. Either \code{"home"} or \code{"away"}. Not passing anything to this param will return both teams.}

\item{limit}{The number of transactions per page to return. The limit and default value is 5000 transactions per page.}

\item{orderBy}{Order transactions by ascending or descending based on transaction id. Default value is \code{"asc"}.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of metrics for a match for each player.
\itemize{
   \item \code{trx.id}  A unique numerical identifier for a given transaction, used for ordering chronologically.
   \item \code{period} The period of the transaction.
   \item \code{period.secs} The elapsed time within the period of a transaction. 
   \item \code{zone.logical} The zone in which the transaction takes place, relative to the squad in possession.
   \item \code{fullname} The full name of the player.
   \item \code{person.id} A unique numerical identifier of the player.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{metric.code} The metric code (ALL_CAPS format).
   \item \code{value} The numeric metric value with no formatting applied.
}
}
\description{
Get time coded transactions for given metric(s) occurring in a match
}
\examples{
getStatisticFlow(216085122)
getStatisticFlow(216085122,metric="TACKLE",period=1,zone='D50')
}
