% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getStatisticFlowExposedFields}
\alias{getStatisticFlowExposedFields}
\title{getStatisticFlowExposedFields}
\format{
An object of class \code{character} of length 10.
}
\usage{
getStatisticFlowExposedFields
}
\description{
This vector contains all of the fields and renames in the getStatisticFlow() function
}
\keyword{internal}
