% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getStatisticFlowWhitelist}
\alias{getStatisticFlowWhitelist}
\title{getStatisticFlowWhitelist}
\format{
An object of class \code{character} of length 10.
}
\usage{
getStatisticFlowWhitelist
}
\description{
This vector contains all of the fields in the getStatisticFlow() function
}
\keyword{internal}
