% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getStoppages}
\alias{getStoppages}
\title{Match Stoppages}
\usage{
getStoppages(matchId, stoppageAttendees = FALSE, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{stoppageAttendees}{A logical value indicating whether to include the attendees of the stoppage. Default is \code{FALSE}.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with all of the stoppages for a given match. 
\itemize{
  \item \code{match.id} A unique numerical identifier for a match.
  \item \code{type.name} The name of the stoppage type (ie. 'Centre Bounce').
  \item \code{type.code} A short code representing the stoppage type.
  \item \code{trx.id} A unique numerical identifier for a transaction.
  \item \code{period} The period in which the transaction occurred.
  \item \code{period.seconds} The elapsed time (in seconds) within the period of the transaction.
  \item \code{stoppage.location.x} The x-coordinate of the stoppage on the field.
  \item \code{stoppage.location.y} The y-coordinate of the stoppage on the field.
  \item \code{stoppage.location.standard.x} The x-coordinate of the stoppage transformed to standard dimensions.
  \item \code{stoppage.location.standard.y} The y-coordinate of the stoppage transformed to standard dimensions.
  \item \code{hitout.id} The transaction id for the hit-out at the stoppage.
  \item \code{hitout.name} The name of the hit-out type.
  \item \code{hitout.code} A short code representing the hit-out type.
  \item \code{hitout.period.seconds} The elapsed time (in seconds) within the period of the hit-out.
  \item \code{hitout.player.displayname} The display name of the player credited with the hit-out.
  \item \code{hitout.player.fullname} The full name of the player credited with the hit-out.
  \item \code{hitout.player.id} A unique numerical identifier for the player credited with the hit-out.
  \item \code{hitout.player.squad.id} A unique numerical identifier for the squad of the player credited with the hit-out.
  \item \code{first.possession.id} The transaction id for the first possession after the stoppage.
  \item \code{first.possession.name} The name of the first possession type.
  \item \code{first.possession.code} A short code representing the first possession type.
  \item \code{first.possession.period.seconds} The elapsed time (in seconds) within the period of the first possession.
  \item \code{first.possession.zone} The zone where the first possession occurred.
  \item \code{first.possession.player.displayname} The display name of the player gaining the first possession.
  \item \code{first.possession.player.fullname} The full name of the player gaining the first possession.
  \item \code{first.possession.player.id} A unique numerical identifier for the player gaining the first possession.
  \item \code{first.possession.player.squad.id} A unique numerical identifier for the squad of the player gaining the first possession.
  \item \code{clearance.id} A unique numerical identifier for the clearance.
  \item \code{clearance.name} The name of the clearance type.
  \item \code{clearance.code} A short code representing the clearance type.
  \item \code{clearance.period.seconds} The elapsed time (in seconds) within the period of the clearance.
  \item \code{clearance.zone} The zone where the clearance occurred.
  \item \code{clearance.player.displayname} The display name of the player performing the clearance.
  \item \code{clearance.player.fullname} The full name of the player performing the clearance.
  \item \code{clearance.player.id} A unique numerical identifier for the player performing the clearance.
  \item \code{clearance.player.squad.id} A unique numerical identifier for the squad of the player performing the clearance.
  \item \code{exit.id} A unique numerical identifier for the stoppage exit.
  \item \code{exit.name} The name of the stoppage exit type.
  \item \code{exit.code} A short code representing the stoppage exit type.
  \item \code{exit.period.seconds} The elapsed time (in seconds) within the period of the stoppage exit.
  \item \code{exit.zone} The zone where the stoppage exit occurred.
  \item \code{exit.type} The type of stoppage exit (e.g., kick, handball).
  \item \code{exit.pressure} The level of defensive pressure applied during the stoppage exit.
  \item \code{exit.player.displayname} The display name of the player performing the stoppage exit.
  \item \code{exit.player.fullname} The full name of the player performing the stoppage exit.
  \item \code{exit.player.id} A unique numerical identifier for the player performing the stoppage exit.
  \item \code{exit.player.squad.id} A unique numerical identifier for the squad of the player performing the stoppage exit.
  \item \code{exit.kicking.foot} The foot used by the player for the stoppage exit kick.
  \item \code{exit.kicking.intent} The intent of the player for the stoppage exit kick.
  \item \code{exit.kicking.distance} The distance of the stoppage exit kick.
  \item \code{exit.kicking.direction} The direction of the stoppage exit kick.
  \item \code{exit.location.x} The x-coordinate of the stoppage exit on the field.
  \item \code{exit.location.y} The y-coordinate of the stoppage exit on the field.
  \item \code{exit.location.x.std} The x-coordinate of the stoppage exit transformed to standard dimensions.
  \item \code{exit.location.y.std} The y-coordinate of the stoppage exit transformed to standard dimensions.
  \item \code{stoppage.attendance.name} The name of the stoppage being attended (ie. 'Ball Up Ruck Contest')
  \item \code{stoppage.attendance.code} A short code representing the type of stoppage being attended.
  \item \code{stoppage.attendance.id} The transaction id that the stoppage attendance is associated to.
  \item \code{stoppage.attendance.period.seconds} The elapsed time (in seconds) within the period of the particular stoppage attendance.
  \item \code{stoppage.attendance.home.fullname1} when \code{stoppageAttendees = TRUE}: The full name of the first player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.displayname1} when \code{stoppageAttendees = TRUE}: The display name of the first player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.personId1} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the first player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.fullname2} when \code{stoppageAttendees = TRUE}: The full name of the second player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.displayname2} when \code{stoppageAttendees = TRUE}: The display name of the second player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.personId2} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the second player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.fullname3} when \code{stoppageAttendees = TRUE}: The full name of the third player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.displayname3} when \code{stoppageAttendees = TRUE}: The display name of the third player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.personId3} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the third player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.fullname4} when \code{stoppageAttendees = TRUE}: The full name of the fourth player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.displayname4} when \code{stoppageAttendees = TRUE}: The display name of the fourth player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.home.personId4} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the fourth player attending the stoppage for the home squad.
  \item \code{stoppage.attendance.away.fullname1} when \code{stoppageAttendees = TRUE}: The full name of the first player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.displayname1} when \code{stoppageAttendees = TRUE}: The display name of the first player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.personId1} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the first player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.fullname2} when \code{stoppageAttendees = TRUE}: The full name of the second player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.displayname2} when \code{stoppageAttendees = TRUE}: The display name of the second player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.personId2} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the second player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.fullname3} when \code{stoppageAttendees = TRUE}: The full name of the third player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.displayname3} when \code{stoppageAttendees = TRUE}: The display name of the third player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.personId3} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the third player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.fullname4} when \code{stoppageAttendees = TRUE}: The full name of the fourth player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.displayname4} when \code{stoppageAttendees = TRUE}: The display name of the fourth player attending the stoppage for the away squad.
  \item \code{stoppage.attendance.away.personId4} when \code{stoppageAttendees = TRUE}: A unique numerical identifier for the fourth player attending the stoppage for the away squad.
}
}
\description{
Get a dataframe of stoppages for a match.
}
\examples{
getStoppages(matchId = 120390401)
}
