% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getStoppages_homeNames}
\alias{getStoppages_homeNames}
\title{getStoppages_homeNames}
\format{
An object of class \code{character} of length 13.
}
\usage{
getStoppages_homeNames
}
\description{
This vector contains all of the column names for the HOME stoppage attendees used when stoppageAttendees = TRUE in getStoppages()
}
\keyword{internal}
