% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getStoppagesExposedFields}
\alias{getStoppagesExposedFields}
\title{getStoppagesExposedFields}
\format{
An object of class \code{character} of length 59.
}
\usage{
getStoppagesExposedFields
}
\description{
This vector contains all of the fields we're exposing when cumulative = FALSE in the getStoppages() function
}
\keyword{internal}
