% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getStoppagesWhitelist}
\alias{getStoppagesWhitelist}
\title{getStoppagesWhitelist}
\format{
An object of class \code{character} of length 65.
}
\usage{
getStoppagesWhitelist
}
\description{
This vector contains all the possible columns in the info portion of the getStoppages() function.
}
\keyword{internal}
