% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{getTotalPages}
\alias{getTotalPages}
\title{getTotalPages}
\usage{
getTotalPages(resp)
}
\value{
EITHER an error message from extractError() or the total number of pages in the API request to iterate over
}
\description{
Helper function used in getStatisticFlow() to stop iteration over pages without warnings printed to console
}
\keyword{internal}
