% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getTRXfile_chainsFields}
\alias{getTRXfile_chainsFields}
\title{getTRXfile_chainsFields}
\format{
An object of class \code{character} of length 12.
}
\usage{
getTRXfile_chainsFields
}
\description{
This vector contains all the fields we're taking when we hit the chains endpoint within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
