% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getTRXfile_twoPlayerStats}
\alias{getTRXfile_twoPlayerStats}
\title{getTRXfile_twoPlayerStats}
\format{
An object of class \code{character} of length 18.
}
\usage{
getTRXfile_twoPlayerStats
}
\description{
This vector contains all the fields we're taking when we pull out the two player stats within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
