% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getTRXfileExposedFields}
\alias{getTRXfileExposedFields}
\title{getTRXfileExposedFields}
\format{
An object of class \code{character} of length 67.
}
\usage{
getTRXfileExposedFields
}
\description{
This vector contains all the fields we're exposing within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
