% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getVenueExposedFields}
\alias{getVenueExposedFields}
\title{getVenueExposedFields}
\format{
An object of class \code{character} of length 13.
}
\usage{
getVenueExposedFields
}
\description{
This vector contains all the fields we're exposing in the getVenueExposedFields function and their equivalent renames.
}
\keyword{internal}
