% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getVenueWhitelist}
\alias{getVenueWhitelist}
\title{getVenueWhitelist}
\format{
An object of class \code{character} of length 13.
}
\usage{
getVenueWhitelist
}
\description{
This vector contains all the possible columns in the getVenue() function.
}
\keyword{internal}
