% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{isAFL}
\alias{isAFL}
\title{isAFL}
\usage{
isAFL(leagueId)
}
\value{
a dataframe (of class "data.frame") or text string (of class "character")
}
\description{
This function is used inside of getSquadSummaryFile() & getPlayerSummaryFile() to run a check on if the matchId passed in is an AFL mens match
}
\keyword{internal}
