% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{matchDetailsCols}
\alias{matchDetailsCols}
\title{matchDetailsCols}
\format{
An object of class \code{character} of length 13.
}
\usage{
matchDetailsCols
}
\description{
This vector contains all of the fields we want to pull from the call to match details in the getSquadSummaryFile() AND getPlayerSummaryFile() functions.
}
\keyword{internal}
