% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{optionalParams}
\alias{optionalParams}
\title{optionalParams}
\usage{
optionalParams(...)
}
\value{
A string with the URL connection and value(s) (ie. "period=1&zone=DM&metric=TACKLE&metric=GOAL&metric=HANDBALL&lastXseconds=500")
}
\description{
This function is used inside of functions with multiple optional arguments that when passed in dictate the URL path (as opposed to functionality within the R function - an example of within-function would be 'stoppageAttendees' param in getStoppages)
optionalParams is used in conjunction with cdAPIresponse() (ie. \code{rawResponse <- cdAPIresponse(endpoint = paste0(baseString, optionalParams("period" = period, "zone" = zone...}
}
\examples{
optionalParams("period" = period, "zone" = zone, "context" = context, "metric" = metric, "team" = team, "lastXseconds" = lastXseconds, "from" = from, "to" = to)
}
\keyword{internal}
