% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{PARAM1_mapping}
\alias{PARAM1_mapping}
\title{PARAM1_mapping}
\format{
An object of class \code{character} of length 26.
}
\usage{
PARAM1_mapping
}
\description{
This vector contains all the mappings for the first parameter within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
