% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{PARAM2_mapping}
\alias{PARAM2_mapping}
\title{PARAM2_mapping}
\format{
An object of class \code{character} of length 14.
}
\usage{
PARAM2_mapping
}
\description{
This vector contains all the mappings for the second parameter within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
