% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{PARAM3_mapping}
\alias{PARAM3_mapping}
\title{PARAM3_mapping}
\format{
An object of class \code{character} of length 7.
}
\usage{
PARAM3_mapping
}
\description{
This vector contains all the mappings for the third parameter within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
