% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{PARAM4_mapping}
\alias{PARAM4_mapping}
\title{PARAM4_mapping}
\format{
An object of class \code{character} of length 5.
}
\usage{
PARAM4_mapping
}
\description{
This vector contains all the mappings for the fourth parameter within the getAFLClubTrxFeed() function and their equivalent renames.
}
\keyword{internal}
