% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{payloadDefaults}
\alias{payloadDefaults}
\title{payloadDefaults (POST Functions Only)}
\usage{
payloadDefaults(payload)
}
\arguments{
\item{payload}{A nested list containing the parcels of data to be returned by the API. See: \code{createPayload()} as a method of simplifying this step.}
}
\value{
A response object from the API call.
}
\description{
This function takes a POST payload as input and will print to the console any parameters of the payload that are missing & signal that defaults for these parameters are being used. For use inside of \code{getPlayerStatsPOST()} & \code{getSquadStatsPOST()} only visable when \code{verbose = TRUE}.
}
\examples{
payloadDefaults(payload = payload)
}
\keyword{internal}
