% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{playerSummaryFile_roundedMetrics1DP}
\alias{playerSummaryFile_roundedMetrics1DP}
\title{playerSummaryFile_roundedMetrics1DP}
\format{
An object of class \code{character} of length 13.
}
\usage{
playerSummaryFile_roundedMetrics1DP
}
\description{
This vector contains the collection of metrics that need to be rounded to 1 decimal place in getPlayerSummaryFile()
}
\keyword{internal}
