% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{playerSummaryFileMetricNames}
\alias{playerSummaryFileMetricNames}
\title{playerSummaryFileMetricNames}
\format{
An object of class \code{character} of length 228.
}
\usage{
playerSummaryFileMetricNames
}
\description{
This vector contains the namees of all the metrics as they appeared in the old Player Summary files available on the support site. This should map 1:1 with squadSummaryFilePlatformMetricNames (used in getSquadSummaryFile())
}
\keyword{internal}
