% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{playerSummaryFilePlatformMetricNames}
\alias{playerSummaryFilePlatformMetricNames}
\title{playerSummaryFilePlatformMetricNames}
\format{
An object of class \code{character} of length 228.
}
\usage{
playerSummaryFilePlatformMetricNames
}
\description{
This vector contains the PLATFORM/API metric names of all the metrics that were in the Player Summary file on the support site. This should map 1:1 with squadSummaryFileMetricNames (used in getSquadSummaryFile())
}
\keyword{internal}
