% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{quietStop}
\alias{quietStop}
\title{quietStop}
\usage{
quietStop()
}
\description{
Helper function used inside of getTotalPages() for getStatisticFlow() to stop iteration over pages without warnings printed to console.
}
\keyword{internal}
