% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{squadSummaryFile_lookup}
\alias{squadSummaryFile_lookup}
\title{squadSummaryFile_lookup}
\format{
An object of class \code{data.frame} with 305 rows and 2 columns.
}
\usage{
squadSummaryFile_lookup
}
\description{
This DF acts as a lookup table containing all the metrics contained in the squad summary file, the names of those metrics as they existed in the old file on the support site (squadSummaryFileMetricNames) and their names now in the API (squadSummaryFilePlatformMetricNames)
}
\keyword{internal}
