% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{squadSummaryFile_roundedMetrics0DP}
\alias{squadSummaryFile_roundedMetrics0DP}
\title{squadSummaryFile_roundedMetrics0DP}
\format{
An object of class \code{character} of length 25.
}
\usage{
squadSummaryFile_roundedMetrics0DP
}
\description{
This vector contains the collection of metrics that need to be rounded to 0 decimal places in getSquadSummaryFile()
}
\keyword{internal}
