% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{squadSummaryFile_roundedMetrics1DP}
\alias{squadSummaryFile_roundedMetrics1DP}
\title{squadSummaryFile_roundedMetrics1DP}
\format{
An object of class \code{character} of length 14.
}
\usage{
squadSummaryFile_roundedMetrics1DP
}
\description{
This vector contains the collection of metrics that need to be rounded to 1 decimal place in getSquadSummaryFile()
}
\keyword{internal}
