% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{squadSummaryFileMetricNames}
\alias{squadSummaryFileMetricNames}
\title{squadSummaryFileMetricNames}
\format{
An object of class \code{character} of length 304.
}
\usage{
squadSummaryFileMetricNames
}
\description{
This vector contains the namees of all the metrics as they appeared in the old Squad Summary files available on the support site. This should map 1:1 with squadSummaryFilePlatformMetricNames (used in getSquadSummaryFile())
}
\keyword{internal}
