% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{squadSummaryFilePlatformMetricNames}
\alias{squadSummaryFilePlatformMetricNames}
\title{squadSummaryFilePlatformMetricNames}
\format{
An object of class \code{character} of length 304.
}
\usage{
squadSummaryFilePlatformMetricNames
}
\description{
This vector contains the PLATFORM/API metric names of all the metrics that were in the Squad Summary file on the support site. This should map 1:1 with squadSummaryFileMetricNames (used in getSquadSummaryFile())
}
\keyword{internal}
