% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{twoplayerstats_empty}
\alias{twoplayerstats_empty}
\title{twoplayerstats_empty}
\format{
An object of class \code{data.frame} with 0 rows and 19 columns.
}
\usage{
twoplayerstats_empty
}
\description{
This dataframe contains all the column names and required class types for the two player stats table, we may need to create an empty DF with these names based on if ruck transactions exist or not
}
\keyword{internal}
