% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{twoplayerstatsbaseNames_trxFile}
\alias{twoplayerstatsbaseNames_trxFile}
\title{twoplayerstatsbaseNames_trxFile}
\format{
An object of class \code{character} of length 57.
}
\usage{
twoplayerstatsbaseNames_trxFile
}
\description{
This vector contains all the column names for the two player stats table, we may need to create an empty DF with these names based on if ruck transactions exist or not
}
\keyword{internal}
