% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{userAgentHeader}
\alias{userAgentHeader}
\title{userAgentHeader}
\format{
An object of class \code{character} of length 1.
}
\usage{
userAgentHeader
}
\description{
The user-agent header used in API request(s) to ID the call has come from the R package, alongside which version number.
}
\keyword{internal}
