% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{zonedMetrics_PlatformNames_Player}
\alias{zonedMetrics_PlatformNames_Player}
\title{zonedMetrics_PlatformNames_Player}
\format{
An object of class \code{character} of length 186.
}
\usage{
zonedMetrics_PlatformNames_Player
}
\description{
The ZONED metrics that go into the POST payload in getPlayerSummaryFile()
}
\keyword{internal}
