% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{zonedMetrics_PlatformNames_Squad}
\alias{zonedMetrics_PlatformNames_Squad}
\title{zonedMetrics_PlatformNames_Squad}
\format{
An object of class \code{character} of length 240.
}
\usage{
zonedMetrics_PlatformNames_Squad
}
\description{
The ZONED metrics that go into the POST payload in getSquadSummaryFile()
}
\keyword{internal}
