% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{zonelessMetrics_PlatformNames_Player}
\alias{zonelessMetrics_PlatformNames_Player}
\title{zonelessMetrics_PlatformNames_Player}
\format{
An object of class \code{character} of length 49.
}
\usage{
zonelessMetrics_PlatformNames_Player
}
\description{
The ZONELESS metrics that go into the POST payload in getPlayerSummaryFile()
}
\keyword{internal}
