% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{zonelessMetrics_PlatformNames_Squad}
\alias{zonelessMetrics_PlatformNames_Squad}
\title{zonelessMetrics_PlatformNames_Squad}
\format{
An object of class \code{character} of length 78.
}
\usage{
zonelessMetrics_PlatformNames_Squad
}
\description{
The ZONELESS metrics that go into the POST payload in getSquadSummaryFile()
}
\keyword{internal}
