% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{zonelessMetricsOther_PlatformNames_Squad}
\alias{zonelessMetricsOther_PlatformNames_Squad}
\title{zonelessMetricsOther_PlatformNames_Squad}
\format{
An object of class \code{character} of length 1.
}
\usage{
zonelessMetricsOther_PlatformNames_Squad
}
\description{
These are metrics that need to be treated differently (ie. Time In Possession essentially gets turned into its own metric) in getSquadSummaryFile()
}
\keyword{internal}
