% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{zoneMapping}
\alias{zoneMapping}
\title{zoneMapping}
\usage{
zoneMapping(coinTossWinner, direction, squadId, period, zonePhysical)
}
\value{
A zone string ie. 'D50', 'DM','F50' etc.
}
\description{
This function is used inside of getAFLClubTrxFeed() to correctly assign zones for transactions based on which way a teams kicking for a given quarter (among other inputs).
}
\keyword{internal}
