% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getChainsExposedFields}
\alias{getChainsExposedFields}
\title{getChainsExposedFields}
\format{
An object of class \code{character} of length 50.
}
\usage{
getChainsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getChains function and their equivalent renames.
}
\keyword{internal}
