% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getFixture}
\alias{getFixture}
\title{Fixture}
\usage{
getFixture(seasonId, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with the fixture for a given season, with one row per match.
\itemize{
   \item \code{competition.code} A short code describing the competition.
   \item \code{season.id} A numerical identifier of a season.
   \item \code{phase.code} A short code describing the phase of the season.
   \item \code{type.id} A numerical indicator of the match type.
   \item \code{type.name} A description of the match type.
   \item \code{round.number} The round number of the match. Continues to count up during finals.
   \item \code{round.phaseNumber} The round number of the match within the phase. Starts again from one for finals.
   \item \code{round.code} A short code representing the name of the round.
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{match.order} The order of a match within a round, sorted by start date and time.
   \item \code{match.start} The start time of a match in UTC time.
   \item \code{match.date} The calendar date of a match in local time.
   \item \code{match.time} The local start time of a match.
   \item \code{home.id} A unique numerical identifier of the home squad.
   \item \code{home.name} The name of the home squad.
   \item \code{home.code} A short code to represent the home squad.
   \item \code{away.id} A unique numerical identifier of the away squad.
   \item \code{away.name} The name of the away squad.
   \item \code{away.code} A short code to represent the away squad.
   \item \code{venue.id} A unique numerical identifier of the match venue.
   \item \code{venue.name} The name of the match venue.
   \item \code{venue.code} A short code to represent the match venue.
   \item \code{status.id} A numerical identifier of the match's current status.
   \item \code{status.name} The match's current status.
   \item \code{status.type.id} A numerical identifier of the match status type.
   \item \code{status.type.name} The match's current status type.
   \item \code{home.goals} The number of goals kicked by the home squad.
   \item \code{home.behinds} The number of behinds kicked by the home squad.
   \item \code{home.points} The number of points scored by the home squad.
   \item \code{away.goals} The number of goals kicked by the away squad.
   \item \code{away.behinds} The number of behinds kicked by the away squad.
   \item \code{away.points} The number of points scored by the away squad.
}
}
\description{
Get the fixture for a given season.
}
\examples{
getFixture(2022)
getFixture()
}
