% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ref.R
\name{getMatchPersons}
\alias{getMatchPersons}
\title{Match Persons}
\usage{
getMatchPersons(matchId, officials = FALSE, ...)
}
\arguments{
\item{matchId}{An unique numerical identifier of a match.}

\item{officials}{Logical. Return match officials? Defaults to FALSE = Return players.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of persons for a match, with one row per person.
\itemize{
   \item \code{match.id} A unique numerical identifier of the match.
   \item \code{squad.id} A unique numerical identifier of the squad. (Players only)
   \item \code{squad.name} The name of the squad. (Players only)
   \item \code{squad.code} A short code to represent the squad. (Players only)
   \item \code{person.id} A unique numerical identifier of a person.
   \item \code{person.name} The person's full name.
   \item \code{person.firstname} The person's first name.
   \item \code{person.surname} The person's surname.
   \item \code{person.display} The person's display name, represented as first initial and surname.
   \item \code{jumper} The number worn on the person's uniform.
   \item \code{selected.id} A numerical identifier of the person's selected position.
   \item \code{selected.name} The person's selected position.
   \item \code{selected.code} A short code representing the peron's selected position.
   \item \code{position.id} A numerical identifier of the person's primary position in that season. (Players only)
   \item \code{position.name} The person's primary position in that season. (Players only)
   \item \code{position.code} A short code representing the peron's primary position in that season. (Players only)
   \item \code{height} The person's height in centimetres. (Players only)
   \item \code{weight} The person's weight in kilograms. (Players only)
   \item \code{DOB} The person's date of birth (YYYY-MM-DD). (Players only)
   \item \code{age} The person's age at the start of the match, to one decimal place. (Players only)
}
}
\description{
Get a list of persons involved in a match. Currently limited to the players or field umpires.
}
\examples{
getMatchPersons(216085122)
getMatchPersons(216085122,officials=TRUE)
}
