% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ref.R
\name{getMetrics}
\alias{getMetrics}
\title{Match Metrics}
\usage{
getMetrics(matchId, ...)
}
\arguments{
\item{matchId}{An unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of metrics for a match.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{id} A unique numerical identifier of a metric.
   \item \code{name} The metric name (singular).
   \item \code{plural} The metric name (plural).
   \item \code{code} The metric code (ALL_CAPS format).
   \item \code{short} An abbreviated metric name (max 25 characters).
   \item \code{level.id} A numerical identifier of the metric level.
   \item \code{level.name} The metric level (Squad, Player, or Both).
   \item \code{format.id} A numerical identifier of the metric format.
   \item \code{format.id} A numerical identifier of the metric format.
   \item \code{description} The text description of the metric definition.
}
}
\description{
Get a list of valid metrics for a match.
}
\examples{
getMetrics(216085122)
}
