% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getPeriodScores}
\alias{getPeriodScores}
\title{Period Scores}
\usage{
getPeriodScores(matchId, cumulative = FALSE, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{cumulative}{Logical. Return scores up to and including that quarter? Defaults to FALSE, which returns scores within the quarter only.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of quarters and scores for each quarter.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{id} The transaction ID of the interchange move, used for ordering chronologically.
   \item \code{home.id} A unique numerical identifier of the home squad.
   \item \code{home.name} The name of the home squad.
   \item \code{home.code} A short code to represent the home squad.
   \item \code{away.id} A unique numerical identifier of the away squad.
   \item \code{away.name} The name of the away squad.
   \item \code{away.code} A short code to represent the away squad.
   \item \code{period} The period of the match.
   \item \code{home.goals} Goals scored by the home squad.
   \item \code{home.behinds} Behinds scored by the home squad.
   \item \code{home.points} Points scored by the home squad.
   \item \code{home.result} Period result for the home squad. W/L/D.
   \item \code{home.margin} Scoreboard margin for the home squad.
   \item \code{away.goals} Goals scored by the away squad.
   \item \code{away.behinds} Behinds scored by the away squad.
   \item \code{away.points} Points scored by the away squad.
   \item \code{away.result} Period result for the away squad. W/L/D.
   \item \code{away.margin} Scoreboard margin for the away squad.
   \item \code{winning.squad.id} The ID of the winning squad.
}
}
\description{
Get a a list of match scores by quarter.
}
\examples{
getPeriodScores(216085122)
getPeriodScores(216085122,cumulative=TRUE)
}
