% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_variables.R
\docType{data}
\name{getPlayerStatsExposedFields}
\alias{getPlayerStatsExposedFields}
\title{getPlayerStatsExposedFields}
\format{
An object of class \code{character} of length 13.
}
\usage{
getPlayerStatsExposedFields
}
\description{
This vector contains all the fields we're exposing in the getPlayerStats function and their equivalent renames.
}
\keyword{internal}
