% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stats.R
\name{getRotations}
\alias{getRotations}
\title{Match Rotations}
\usage{
getRotations(matchId, ...)
}
\arguments{
\item{matchId}{A unique numerical identifier of a match.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of rotations for the match, with one row per interchange move.
\itemize{
   \item \code{match.id} A unique numerical identifier of a match.
   \item \code{id} The transaction ID of the interchange move, used for ordering chronologically.
   \item \code{squad.id} A unique numerical identifier of the squad.
   \item \code{squad.name} The name of the squad.
   \item \code{squad.code} A short code to represent the squad.
   \item \code{period} The period of the interchange move.
   \item \code{secs} The elapsed time within the period of the interchange move.
   \item \code{off.id} A unique numerical identifier of the player coming off the ground.
   \item \code{off.name} The full name of the player coming off the ground.
   \item \code{off.display} The display name of the player coming off the ground, represented as first initial and surname.
   \item \code{off.reason} The reason for the interchange move.
   \item \code{off.code} A short text description of the reason for the interchange move.
   \item \code{on.id} A unique numerical identifier of the player coming on the ground. Empty for start-of-match initialisation of the interchange bench.
   \item \code{on.name} The full name of the player coming on the ground. Empty for start-of-match initialisation of the interchange bench.
   \item \code{on.display} The display name of the player coming on the ground, represented as first initial and surname. Empty for start-of-match initialisation of the interchange bench.
}
}
\description{
Get a list of rotations for a match.
}
\examples{
getRotations(216085122)
}
