% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ref.R
\name{getSquads}
\alias{getSquads}
\title{Squad List}
\usage{
getSquads(seasonId, leagueId = 1, levelId = 1, ...)
}
\arguments{
\item{seasonId}{A numerical identifier of a season. If missing, defaults to the current season of the League Level defined by 'leagueId' and 'levelId'.}

\item{leagueId}{A unique numerical identifier of a league. Defaults to AFL Men's Premiership.}

\item{levelId}{A unique numerical identifier of a level. Defaults to Seniors.}

\item{...}{Arguments to be passed to internal functions, such as \code{envir} or \code{version}.}
}
\value{
A data frame with a list of squads, with one row per squad.
\itemize{
   \item \code{season.id} A numerical identifier of a season.
   \item \code{id} A unique numerical identifier of the squad.
   \item \code{name} The name of the squad.
   \item \code{code} A short code to represent the squad.
}
}
\description{
Get a list of squads participating in a season.
}
\examples{
getSquads(2022)
getSquads()
}
