% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_fields.R
\docType{data}
\name{getSquadStatsPOSTExposedFields}
\alias{getSquadStatsPOSTExposedFields}
\title{getSquadStatsPOSTExposedFields}
\format{
An object of class \code{character} of length 10.
}
\usage{
getSquadStatsPOSTExposedFields
}
\description{
This vector contains all the fields we're exposing in the getSquadStatsPOST function (not including info) and their equivalent renames.
}
\keyword{internal}
