% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{to_minsec}
\alias{to_minsec}
\title{to_minsec}
\usage{
to_minsec(x, leadingZero = FALSE)
}
\arguments{
\item{x}{Seconds, as an integer.}

\item{leadingZero}{Logical. Add a leading zero to minute values less than 10?}
}
\value{
A text string in the form "MI:SS" - eg. "9:54" if @param leadingZero is FALSE, "09:54" if @param leadingZero is TRUE.
}
\description{
Convert an integer of seconds elapsed to a text string with minutes and seconds.
}
\examples{
to_minsec(594)
}
